+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.1] - TBD
 
 ### Fixed
-- Corrected bug in how summation is occuring in sort_salience_map()
+- Corrected bug in how summation is occuring in sort_salience_map() and added unit test